"On-the-fly" high-level account allocation and releasing.
If args.hlvl_workers is set to "0" but encounters are enabled, the Account Manager won't keep the accounts allocated.
This means that the queue of high-level accounts won't hold the account after it is used.
If args.hlvl_kph is set to "0", the speed limit won't be checked and we can maximize the number of encounters an account can do.