Improvements in permanent ban detection.
Since accounts are shared between instances we needed a way to keep track of the number of failed login attempts.
The solution we had was local/runtime only and wouldn't work for "on-the-fly" account allocation.
Renamed Account model field 'fail' to 'failed' and changed its type to SmallInt in order to have a persistent counter of login failures. 'fail' field was only used to simbolize account was on failed account pool, it had almost no purpose, now it's an important field.
Added '--account-max-failures' to control the number of consecutive login attempts failed it takes to flag the account as perm banned. Default is 5, meaning the account will have to fail N login-retries, on 5 different occasions to be considered banned. Once it fails the first time account will be put to sleep for a small period of time and will attempt to login again later.
Reverted some changes in '--hlvl-workers-max', it is only used when '--hlvl-workers' is 0 (on-the-fly allocation).